+++ src/Core/src/Handlers/SwipeItemMenuItem/SwipeItemMenuItemHandler.Standard.cs
