+++ src/Core/src/Handlers/SwipeItemMenuItem/SwipeItemMenuItemHandler.cs
