+++ src/Core/src/Handlers/SwipeItemView/ISwipeItemViewHandler.cs
@@ -1,16 +1,18 @@
-﻿#if __IOS__ || MACCATALYST
+﻿#if IOS || MACCATALYST
 using PlatformView = Microsoft.Maui.Platform.ContentView;
 #elif MONOANDROID
 using PlatformView = Microsoft.Maui.Platform.ContentViewGroup;
+#elif WINDOWS
+using PlatformView = Microsoft.UI.Xaml.FrameworkElement;
+#elif NETSTANDARD || (NET6_0 && !IOS && !ANDROID)
+using PlatformView = System.Object;
 #endif
 
 namespace Microsoft.Maui.Handlers
 {
-#if __IOS__ || MACCATALYST || MONOANDROID
 	public interface ISwipeItemViewHandler : IViewHandler
 	{
 		new ISwipeItemView VirtualView { get; }
 		new PlatformView PlatformView { get; }
 	}
-#endif
 }
