+++ src/Core/src/Handlers/SwipeItemView/SwipeItemViewHandler.Windows.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using Microsoft.UI.Xaml;
+
+namespace Microsoft.Maui.Handlers
+{
+	public partial class SwipeItemViewHandler : ViewHandler<ISwipeItemView, FrameworkElement>, ISwipeItemViewHandler
+	{
+		protected override FrameworkElement CreatePlatformView()
+		{
+			throw new NotImplementedException();
+		}
+
+		public static void MapContent(ISwipeItemViewHandler handler, ISwipeItemView page)
+		{
+		}
+
+		public static void MapVisibility(ISwipeItemViewHandler handler, ISwipeItemView view)
+		{
+		}
+	}
+}
