+++ src/Core/src/Handlers/SwipeItemView/SwipeItemViewHandler.cs
@@ -1,0 +1,43 @@
+﻿using System;
+#if IOS || MACCATALYST
+using PlatformView = Microsoft.Maui.Platform.ContentView;
+#elif MONOANDROID
+using PlatformView = Microsoft.Maui.Platform.ContentViewGroup;
+#elif WINDOWS
+using PlatformView = Microsoft.UI.Xaml.FrameworkElement;
+#elif NETSTANDARD || (NET6_0 && !IOS && !ANDROID)
+using PlatformView = System.Object;
+#endif
+
+namespace Microsoft.Maui.Handlers
+{
+	public partial class SwipeItemViewHandler : ViewHandler<ISwipeItemView, PlatformView>, ISwipeItemViewHandler
+	{
+		public static IPropertyMapper<ISwipeItemView, ISwipeItemViewHandler> Mapper = new PropertyMapper<ISwipeItemView, ISwipeItemViewHandler>(ViewHandler.ViewMapper)
+		{
+			[nameof(ISwipeItemView.Content)] = MapContent,
+			[nameof(ISwipeItemView.Visibility)] = MapVisibility
+		};
+
+		public static CommandMapper<ISwipeItemView, ISwipeItemViewHandler> CommandMapper = new(ViewHandler.ViewCommandMapper)
+		{
+		};
+
+		public SwipeItemViewHandler() : base(Mapper, CommandMapper)
+		{
+		}
+
+		protected SwipeItemViewHandler(IPropertyMapper mapper, CommandMapper? commandMapper = null)
+			: base(mapper, commandMapper ?? CommandMapper)
+		{
+		}
+
+		public SwipeItemViewHandler(IPropertyMapper? mapper = null) : base(mapper ?? Mapper)
+		{
+		}
+
+		ISwipeItemView ISwipeItemViewHandler.VirtualView => VirtualView;
+
+		PlatformView ISwipeItemViewHandler.PlatformView => PlatformView;
+	}
+}
